Initialise member data in constructors rather than header files. Move function definitions to serparate cpp files rather than defining in header. Remove unnecessary includes.
